+++ handler/producer/produce.go
@@ -7,11 +7,11 @@ import (
 var (
 	_ Roundtrip = Proxies{}
 	_ Roundtrip = Requests{}
-	_ Roundtrip = Sequences{}
+	_ Roundtrip = Parallel{}
 	_ Roundtrip = Sequence{}
 )
 
 type Roundtrip interface {
-	Produce(req *http.Request, results chan<- *Result)
+	Produce(req *http.Request) chan *Result
 	Len() int
 }